server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "FastAPI server for Knowledge Graph API"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[project.scripts]
dev = "uvicorn src.main:app --reload --port 3334"
